Scripts and results based on AFF2024
We predicted the conformational dynamics of CAPs based on AFF2024 and present the results. Due to upload capacity limitations, only a portion of the results has been uploaded. The pipeline of AFF2024, which is based on AlphaFold2, appears to inadequately simulate the conformational dynamics of CAPs. Despite multiple attempts, we were unable to achieve satisfactory results. We only list the results of one experimental run here. Consequently, we attempted to use a method based on AlphaFold3, incorporating energetic frustration properties to predict the conformational dynamics of CAPs.